Move left pane within datascape base route since we do not need it for query builder
