+++ mathesar_ui/src/component-library/tree/Tree.scss
@@ -1,7 +1,6 @@
 .tree {
   > .search-box {
     padding: 0px 15px 10px;
-    background: #f9fafdcc;
   }
 
   > .empty {
